+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.0.1] - 2026-04-20
+
+### Changed
+
+- VectorHeader.zero_point changed from i32 to i16
+- VectorHeader.padding: u16 removed
+- index: removed duplicate use std::sync::LazyLock;
+
 ## [3.0.0] - 2026-04-19
 
 ### Added